Add couple of extra utilities - dd and fstype.
dd supports most of POSIX bits except for block/unblock conversions.
fstype detects the type and size of the filesystem passed on its stdin,
emitting a couple of expressions suitable for use in eval to set shell
variables.

This allows us to automatically detect the filesystem type when
rootfstype= isn't specified, and when we're loading a ramdisk (via
my load_ramdisk script) we can work out if we need to gunzip it
first, and in the case that we don't, we know how many 1K blocks
to ask dd to copy for us.