[PATCH] avoid name clash on nfs_fh in nfsmount
Avoid a name clash in nfsmount on Fedora glibc.

The name nfs_fh is used in both kernel and in nfsmount
application, but one uses it to mean the argument to mount(2),
the other means the data as captured from the wire.

The package glibc-kernelheaders in Fedora makes the kernel
structure visible to user space, so conflict results
in compilation errors.  With klibc, the homonym is harmless
because its hidden by ifdef __KERNEL__.

Signed-off-by: Erik van Konijnenburg <ekonijn@xs4all.nl>
Signed-off-by: H. Peter Anvin <hpa@zytor.com>