kbuild: Included kinit, nfsmount and ipconfig in kernel build
The kinit (and the rest) are build both as static and shared linked
executables. To support this in a clean way the support in
Kbuild.klibc was reworked considerably.
Now to tell when you want a static linked executables use:
static-y := binary

Likewise for shared executables use:
shared-y := binary

To test things out a Kbuild file was created for kinit, ipconfig and nfsmount.
Things worked out pretty well and the syntax is kept nice and readable.
Due to the fact that kinit uses ipconfig and nfsmount they were in the
kernel placed as subdirectories for kinit. This is the only sane way to
express the dependencies when doing recursive build like in the kernel.

With the current scheme a .o file is created for ipconfig and nfsmount.
If we are better served with a ar file then this is easy to fix.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>