kbuild: add Kbuild for utils and ash
Added a Kbuild file for utils/ and ash.
For utils/ I had to tell kbuild for each single binary
what .o file it used because the binary is placed in a subdirectory.

For ash I had to teach some of the script files that the output files
are stored in a directory != current directory.

cp-to-kernel also copies utils and ash now.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>