kbuild: let static-y + shared-y depend on lib.a
Introducing this dependency allows us to build a lib.a in a directory
using: lib-y := file.o
And then using EXTRA_USERLDFLAGS we can tell the linker to use this file.

There is no users in klibc atm, but the possibility was so good that
I decided to keep it (developed it for utils/ but decided for a better
solution there).

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>