kbuild: fix cleaning in kernel tree
To support cleaning a bit of resturcturing was needed in usr/Kbuild.
The current structure descend into klibc binary subdirectories only
if CONFIG_KLIBC is selected.

The is not yet support for adding the binaries to the initramfs image.
That will wait until we find a nice way to do so.

Also added a few statements to various Kbuild files to let 'make clean'
pick up all output files.
There is room for some generalization - but that will wait until klibc hit mainline.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>