kbuild: s/USER/KLIBC/g
In general replaced the term USER with the more specific term KLIBC.
When included in the kernel the term USER is ambigious,
especially if kbuild are expanded with specific UML support.

So introduce the common prefix 'KLIBC' as replacement for USER.

While in the process Kbuild.klibc is now more careful to pick up
definitions from arch specific MCONFIG.
This is made backward compatible since we share the arch specific
MCONFIG files with the old klibc buildsystem for now.
Also brought back EMAIN - we may need it soon for klcc.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>