kbuild: move klcc to directory named klcc
In a preparation to kbuildtify klibc move klcc down to a klcc subdirectory.
kbuild is used to build klcc, so started to add some kbuild definitions
to the top-level Makefile.

In klinc.config the old names are kept on the left-hand side. This was chosen
so we do not break compatibility with existing klcc users.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>