Non-kbuild klibc obsoleted
On Tue, Dec 13, 2005 at 09:48:44AM -0800, H. Peter Anvin wrote:
> Hi there,
>
> Just found a problem... the old Makefile produced e.g.
> utils/{static,shared,static.g,shared.g} with and without debugging
> turned on.  This is important, since with klibc sometimes these aren't
> just the same thing as building with debugging and then stripping; that
> doesn't work too well on e.g. x86-64.
>
> Any idea how to get this back?

I took the simple approch and always do it when building
the shared / static binaries.
I assume this cover your need.

Please note one major drawback. If you delete the *.g file kbuild
will NOT notice and will not rebuild it.

	Sam

kbuild: create stripped versions of programs

When building utilities always create a file.g that is the non-stripped
binary. Please note that the file.g file is created as a side effect,
so kbuild does not know about it. In other words kbuild will not rebuild
the .g file if you delete it. And it will overwrite it no matter the
timestamp on the file

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>