alpha grab errno from right register after syscall
The next problem in line was that when trying to clear the initramfs to
free memory before overmounting, run-init was getting EPERM instead of
EISDIR when calling unlink() on directories.  After stepping through both
glibc and klibc linked binaries, I've concluded that the issue is simply
that klibc is trying to read the errno from the wrong register.

Initially, I was concerned this meant unlink was different from other
syscalls on alpha, which didn't make sense; then I realized that none of
the other syscalls I was looking at were *returning* a non-zero errno,
just a return value.

verified by indepent test boots that both patches fixes alpha klibc.

Signed-off-by: maximilian attems <janitor@sternwelten.at>
Signed-off-by: H. Peter Anvin <hpa@zytor.com>