klibc: fix sparc build
Fix build of klibc on sparc. Following patch is only tested on a kernel
build but should work on a native klibc build too.

The patch fixes a bug in klibc/Kbuild so assignments to the variable
targets in arch/$(KLIBCARCH)/Makefile.inc is not lost.
Without this 'make clean' did not work as expected.

Based on input from maximilian attems <maks@sternwelten.at> and
Fabio Massimo Di Nitto <fabbione@ubuntu.com>.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>