[klibc] Detect the sizes of various types, and make available to sysstub.ph.
This additional code effectively queries the C compiler for the sizes of
various types, and makes an associative array %typesize available to
sysstub.ph.  This is currently not used, but it's expected that some
architectures, e.g. s390, will need this to determine which registers
go where, and how many registers are needed.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>