[klibc] ARM: More EABI and Thumb fixes.  Make the syscall stubs smaller.
More work on getting EABI and Thumb to reliably build.  It should all
be good now.

When building for EABI or Thumb, make the syscall stubs extra small by
using "bl" followed by an immediate; since we have already saved lr
the additional clobber of lr is not a problem.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>