[klibc] Support Perl 5.6; add test for type signedness
There seems to be a consensus that Perl 5.6 should be supported,
which means we can't use the actually sane 3-operand form of open().
Furthermore, unpack("Z*") appears to be broken in Perl 5.6.1
-- at least in the version I built -- so open-code the decoding
of the typesize file.

Finally, while we're changing this code around, add a test for the
signedness of types.  This may end up being very useful for cases
where registers need to be sign- or zero-extended, as appropriate.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>