Temporary fix for the kinit parallel build problem
How to deal with the kinit stuff...

On Thu, Jul 27, 2006 at 01:33:18PM -0700, H. Peter Anvin wrote:
> Hi Sam,
>
> Any thoughts about how to deal with the kinit stuff?

Here is a quick hack.

The way I try to solve it is:

1) define a list of targets that must be build before descending.
   This is the "common/" stuff.
2) In resume/ define a number of .o files that shall be linked but
   not know as target by make.

Naming of variables are bad and they need better integration (kbuild
shall do objectify, assign targets etc.).
This was just something so you had a possibility to toy with it while
I'm having fun.

	Sam