sh: fix build
Hi Folks

A few trivial changes to the sh arch support to get everything to build.

Everything links OK but although we can welcome the world with open arms
few of the other test program run correctly ...

--
Daniel Thompson (STMicroelectronics) <daniel.thompson@st.com>
1000 Aztec West, Almondsbury, Bristol, BS32 4SQ. 01454 462659

If a car is a horseless carriage then is a motorcycle a horseless horse?