[klibc] [EXPAND] Removed herefd hack
The herefd hack goes back more than a decade.  it limits the amount of
memory we have to allocate when expanding here-documents by writing the
result out from time to time.  However, it's no longer safe because the
stack is used to place intermediate results too and there we certainly
don't want to write them out should we be short on memory.

In any case, with today's computers we can afford to keep the entire
result in memory and write them out at the end.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: maximilian attems <max@stro.at>