[klibc] [EXPAND] Fix slash treatment in expmeta
The change

   [EXPAND] Do not quote back slashes in parameter expansions outside quotes

triggered a latent bug in expmeta where the forward slashes when preceded
by a blackslash weren't recognised as directory separators.  This was hidden
because a work-around was put in place for glob(3) which meant that we never
had any backslashes immediately before forward slashes.

This patch fixes the metaflag loop to recognise forward slashes even when
they follow a backslash.

Thanks to Daniel Hahler for reporting this problem.

Test case:

	echo "/"root*

Old result:

	/root*

New result:

	/root

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: maximilian attems <max@stro.at>