[klibc] ipconfig may discard useful packets
In main.c, function loop(), there's a call to packet_discard().
Unfortunately, there was a logic error in the return values of
some functions, which caused it either to
1. Be called when a useful packet was in the queue,
   such as DHCPACK, and so this packet was discarded, a
   timeout occured and the correct address was acknowledged
   only by a retransmittion after a few seconds, or
2. Be called when no packets where in the queue, so
   packet_discard() just delayed until some new packet
   arrived, and this new packet was again discarded.

I've seen many times ipconfig needing many seconds to
configure a card (in my school we have 2 dhcp servers, maybe
this makes the problem more common).

I think I corrected the function return values, and in
my tests (with either one or two dhcp servers) now ipconfig
gets an address in only a few msec.
Sorry, but this means that many functions were modified,
but in the good side, the return values were commented.
I did my best to avoid any regressions, but a second look
at the code or extra testing may be needed.

A scenario where the bug occurs:
In main.c:
	nr = do_pkt_recv(pkt_fd, now.tv_sec);
	if (nr == 1)
		break;
	else if (nr == 0)
		packet_discard(); <== This one

The following functions may be called:
do_pkt_recv > process_receive_event > e.g. dhcp_recv_offer > dhcp_recv
> packet_recv

packet_recv may receive a not-appropriate packet, and jump to
discard_pkt, in which case it *discards* the packet and returns 0.
This return value (0) will propagate to main, so (nr == 0) and
packet_discard() will be called *again*, causing the NIC to wait until
it happens to
receive some packet.

Signed-off-by: Άλκης Γεωργόπουλος (alkisg@gmail.com)
Signed-off-by: maximilian attems <max@stro.at>