[klibc] add getmntent()
internaly use reentrant getmntent_r(), can be exported on need.
take care to use reentrant strtok_r() functions too.
Thanks to Karel Zak <kzak@redhat.com> for pointing that out.

getmntent is used in several places to check that a news fs
won't be created on an already mounted partition.
it is also needed for mount(8) to list mounted fs.

Signed-off-by: maximilian attems <max@stro.at>