[klibc] [EVAL] Pass EV_TESTED into evalcmd
This patch fixes the case where the eval command is used with
set -e and as part of a construct that should not cause the
shell to abort, e.g., as part of the condition of an if statement.

This is achieved by propagating the EV_TESTED flag into the
evalstring function through evalcmd.  As this alters the prototype
of evalcmd it is now invoked explicitly by evalbltin.  The built-in
infrastructure has been changed to accomodate this special case.

In order to ensure that the EXIT trap is properly executed this
patch clears evalskip in exitshell.  This wasn't needed before
because of the broken way evalstring worked where it always clears
evalskip when called by minusc.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: maximilian attems <max@stro.at>