[klibc] fix build on native i386
When trying to build klibc on a native i386 box it
failed to include asm/posix_types.h

This was tracked down to a wrong include path.
For x86_64 it was correct.
Fix it for i386.

This should fix the build problem reported Phil Howard.

"FYI basically same patch was done and tested in OpenEmbedded on 2010-07-01"
- Andrea Adami <andrea.adami@gmail.com>

6f08923d3024f108b6262258c913573fb6ea78e6 fixed allmost all other unified
archs.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>
Cc: Phil Howard <kunnskaperen@gmail.com>
Signed-off-by: maximilian attems <max@stro.at>