[klibc] ipconfig: Only peek and discard packets from specified device.
This patch fixes a bug on systems with multiple connected network devices.
As packet_peek uses all devices to receive data instead of a specific
device. As the return value was never reset it was possible that packets
from other devices were returned by packet_peek. That means that the
ifindex did not match any ifindex of the specified devices the packet was
never removed and packets for the correct device were never processed.

This patch enhance packet_peek and packet_discard to only work on packages
for the specified device instead of all packets.

Signed-off-by: Ulrich Dangel <uli@spamt.net>
Signed-off-by: maximilian attems <max@stro.at>