[klibc] ipconfig: Don't try to guess the nameserver
klibcs ipconfig tries to guess the nameserver if no nameserver is
provided. This may happen due to misconfigured dhcp or the use of
the ip= parameter.

This patch removes the guessing from ipconfig as it is not possible for
any other script to determine if the provided nameserver is valid or
a guessed one. If the old behavior is really needed this could be easily
implemented in an external script.

Analysis shows that  ip_nameserver is unused outside of ipconfig,
nor any sign of gethostbyname() or such. mount.c doesn't do dns resolving,
so I'd guess your patch is a welcome fix. I haven't seen this guessing in
linux >= 2.0.

Closes: #594638

Signed-off-by: Ulrich Dangel <uli@spamt.net>
Reviewed-by: H. Peter Anvin <hpa@zytor.com>
Signed-off-by: maximilian attems <max@stro.at>