[klibc] s390x: klcc searches for include in the wrong path
s390 and s390x are sharing the same include files, in
/usr/lib/klibc/include/arch/s390. However klcc seatches for the s390x
includes in /usr/lib/klibc/include/arch/s390x, which doesn't exist.

The patch below fixes the issue by using the ARCHDIR variable instead
of the ARCH one in klcc. It's one possible fix, another possible fix
would be to use a symlink, or to split s390 and s390x in two different
architectures as anyway they don't share so much code.

Signed-off-by: maximilian attems <max@stro.at>